seat-map-omarchy: instrument the connect path the way launch is instrumented
The output flow blocks and records network requests by wrapping chromium.launch
and routing every context. A verify driven over --connect returns a Browser from
chromium.connect and goes through the same newContext calls, so the wrapper
simply never saw it: an empty receipt, an instrument silently seeing nothing.
Both entry points now share one instrument function, and the previous label's
verify legs reach real hardware over a browser server the same way the
compatibility flow's do.

The server helpers move to their own module, since two flows now use them, and
verifyArgs takes the directory to verify so the output flow's served argument
fits. Its arm covers that directory reaching both labels alike, because the
label difference has to survive a verify that is not run at the app root.
